second commit

+++ 04-spring-boot-rest-crud/spring-boot-rest-crud-employee/src/main/java/com/luv2code/springboot/cruddemo/rest/EmployeeRestController.java
@@ -44,7 +44,13 @@ public class EmployeeRestController {
     }
 
     @DeleteMapping("/employees/{id}")
-    public void deleteEmployee(@PathVariable int id) {
+    public String deleteEmployee(@PathVariable int id) {
+        Employee employee = employeeService.findById(id);
+
+        if(employee == null)
+            throw new RuntimeException("Employee id not found - " + id);
+
         employeeService.deleteById(id);
+        return "Deleted employee id - " + id;
     }
 }
